Update DependencyStatus2App.html
This app is somewhat painful to utilize day-after-day. I've updated this branch to only display the current Iteration, and I've expanded the box to span the width and height of the page. The commented out defect portion is also included in this version. 

It'd be better if this single display was driven by the Iteration selector, however, I wasn't sure how to achieve this. Instead, the iteration query is edited to as follows: 

 query : "((StartDate <= today) AND (EndDate >= today))",

It'd be nice if someone could edit this to populate the query from the Iteration selector to display whatever inputs the user selects, but I've spent too much time on this already. Cheers. :)